更新操作

+++ application/admin/controller/Admin.php
@@ -8,6 +8,8 @@ use think\Db;
 use think\facade\Request;
 use app\admin\validate\Admin as AdminValidate;
 use app\admin\model\Admin as AdminModel;
+use think\facade\Session;
+
 class Admin extends Base
 {
     /**
@@ -15,12 +17,20 @@ class Admin extends Base
      */
     public function list()
     {
-        //查询所有管理信息
-        $info = Db::name('admin')->order('create_time','desc')->paginate(10);
+        //获取当前管理员ID
+        $id = Session::get('Admin.id');
+        //如果当前管理员为超级管理员，则输出全部管理员信息
+        if ($id == 1){
+            //查询所有管理信息
+            $info = Db::name('admin')->order('create_time','desc')->paginate(10);
+        }else{
+            //查询当前管理员的信息
+            $info = Db::name('admin')->where('id',$id)->order('create_time','desc')->paginate(10);
+        }
         //给模板赋值
         $this->assign(['admin'=>$info]);
         //渲染模板
-        return $this->fetch('list');
+        return $this->fetch('admin/list');
     }
 
     /*
@@ -32,7 +42,7 @@ class Admin extends Base
         if (request()->isAjax()){
             //接收前台传过来的数据
             $data = Request::param();
-            //验证数据
+            //验证数据格式
             $validate = new AdminValidate;
             if (!$validate->sceneAdd()->check($data)){
                 $this->error($validate->getError());
@@ -63,7 +73,7 @@ class Admin extends Base
         $info = Db::name('group')->where('id','>',1)->field('id,name')->select();
         //给模板赋值
         $this->assign(['add'=>$info]);
-        return $this->fetch('add');
+        return $this->fetch('admin/add');
     }
 
     /**
@@ -87,27 +97,47 @@ class Admin extends Base
             if (!$validate->sceneEdit()->check($data)){
                 $this->error($validate->getError());
             }
-            //实例化对象
-            $admin = new AdminModel();
-            //执行更新并过滤非数据表字段
-            $res = $admin->allowField(true)->save($data,['id'=>$data['id']]);
-            if ($res){
-                //同时更新中间表
-                Db::name('group_access')->where('uid',$data['id'])->update(['group_id'=>$data['group']]);
-                //记录日志
-                $this->log("修改了管理员：{$info['user']}的个人信息！");
-                $this->success("更新成功！",'Admin/list');
+            //超级管理员和当前管理员（自己）的状态不能修改
+            if ($data['id'] == 1 && $data['status'] == 0){
+                $this->error("超级管理员状态不能修改！");
+            }elseif($data['id'] == Session::get('Admin.id') && $data['status'] == 0){
+                $this->error("自己的状态不能修改！");
             }else{
-                //记录日志
-                $this->log("修改管理员：{$data['user']}的个人信息失败！");
-                $this->error("更新失败！");
+                //实例化对象
+                $admin = new AdminModel();
+                //如果为超级管理员，则可以修改密码，否则不行
+                if (Session::get('Admin.id') == 1){
+                    //判断密码是否已经修改,则重新加密
+                    if ($data['password'] !== ""){
+                        $data['password'] = password_hash($data['password'],PASSWORD_BCRYPT);
+                    }else{
+                        //删除传过来的空字符串密码
+                        unset($data['password']);
+                    }
+                    //执行更新并过滤非数据表字段
+                    $res = $admin->allowField(true)->save($data,['id'=>$data['id']]);
+                }else{
+                    //执行更新并过滤非数据表字段
+                    $res = $admin->allowField(true)->save($data,['id'=>$data['id']]);
+                }
+                if ($res){
+                    //同时更新中间表
+                    Db::name('group_access')->where('uid',$data['id'])->update(['group_id'=>$data['group']]);
+                    //记录日志
+                    $this->log("修改了管理员：{$info['user']}的个人信息！");
+                    $this->success("更新成功！",'Admin/list');
+                }else{
+                    //记录日志
+                    $this->log("修改管理员：{$data['user']}的个人信息失败！");
+                    $this->error("更新失败！");
+                }
             }
         }
         //查询权限组数据表所有信息
         $group = Db::name('group')->field('id,name')->select();
         //给模板赋值
         $this->assign(['admin'=>$info,'group'=>$group]);
-        return $this->fetch('edit');
+        return $this->fetch('admin/edit');
     }
 
     /**
@@ -118,14 +148,26 @@ class Admin extends Base
         if (request()->isAjax()){
             //接收前台传过来的ID
             $id = Request::param('id');
-            //进行删除操作
-            $res = Db::name('admin')->delete($id);
-            if ($res){
-                //同时删除中间表关联
-                Db::name('group_access')->where('uid',$id)->delete();
-                $this->success("删除成功！");
+            //转为数组
+            $array = explode(',',$id);
+            //判断是否存在超级管理员，是则不能删除
+            if (in_array(1,$array) == false){
+                //判断是否存在自己,是则不能删除
+                if (in_array(Session::get('Admin.id'),$array) == false){
+                    //进行删除操作
+                    $res = Db::name('admin')->delete($id);
+                    if ($res){
+                        //同时删除中间表关联
+                        Db::name('group_access')->where('uid',$id)->delete();
+                        $this->success("删除成功！");
+                    }else{
+                        $this->error("删除失败！");
+                    }
+                }else{
+                    $this->error("自己不能删除！");
+                }
             }else{
-                $this->error("删除失败！");
+                $this->error("超级管理员不能可删除！");
             }
         }
     }
