+++ application/admin/controller/Group.php
@@ -10,6 +10,7 @@ use app\admin\model\Admin;
 use app\admin\model\Group as GroupModel;
 use think\facade\Request;
 use app\admin\validate\Group as GroupValidate;
+use think\facade\Session;
 
 class Group extends Base
 {
@@ -18,8 +19,17 @@ class Group extends Base
      */
     public function list()
     {
-        //查询所有权限组信息
-        $info = GroupModel::order('create_time','desc')->paginate(10);
+        //获取当前管理员ID
+        $id = Session::get('Admin.id');
+        //如果当前为超级管理员，则输出全部权限组信息
+        if ($id == 1){
+            //查询所有权限组信息
+            $info = GroupModel::order('create_time','desc')->paginate(10);
+        }else{
+            //只查询当前管理员权限组的权限信息
+            $gid = Db::name('group_access')->where('uid',$id)->field('group_id')->find();
+            $info = GroupModel::where('id',$gid['group_id'])->order('create_time','desc')->paginate(10);
+        }
         foreach ($info as $key=>$val){
             //关联查询
             $user = Db::view('group_access','uid,group_id')
@@ -30,7 +40,7 @@ class Group extends Base
         }
         //给模板赋值
         $this->assign(['group'=>$info]);
-        return $this->fetch('list');
+        return $this->fetch('group/list');
     }
 
     /**
@@ -67,7 +77,7 @@ class Group extends Base
             $one[$key]['two'] = $two;
             //循环二级数组
             foreach ($one[$key]['two'] as $item => $value){
-                $three = Db::name('rule')->where('pid',$value['id'])->field(['id','name,pid as ppid'])->select();
+                $three = Db::name('rule')->where('pid',$value['id'])->field(['id','name','pid as ppid'])->select();
                 //循环将pid赋值给$three
                 for ($i=0; $i<count($three); $i++){
                     $three[$i]['pid'] = $value['pid'];
@@ -76,7 +86,7 @@ class Group extends Base
             }
         }
         $this -> assign(['add'=>$one]);
-        return $this->fetch('add');
+        return $this->fetch('group/add');
     }
 
     /**
@@ -93,6 +103,12 @@ class Group extends Base
             if (!$validate->sceneEdit()->check($data)){
                 //验证不通过输出提示信息
                 $this->error($validate->getError());
+            }
+            //超级权限组和正在使用的权限组状态不能修改
+            if ($data['id'] == 1 && $data['status'] == 0){
+                $this->error("超级权限组的状态不能修改！");
+            }elseif($data['id'] == Session::get('Admin.id') && $data['status'] == 0){
+                $this->error("正在使用的权限组状态不能修改！");
             }
             //实例化对象
             $group = new GroupModel();
@@ -136,7 +152,7 @@ class Group extends Base
         }
         //赋值变量给模板
         $this -> assign(['edit'=>$one,'rules'=>$info]);
-        return $this->fetch('edit');
+        return $this->fetch('group/edit');
     }
 
     /**
@@ -148,12 +164,23 @@ class Group extends Base
         if (request()->isAjax()){
             //接收前台传过来的ID
             $id = Request::param('id');
-            //进行删除操作
-            $res = Db::name('group')->delete($id);
-            if ($res){
-                $this->success("删除成功！");
+            //转为数组
+            $array = explode(',',$id);
+            //判断是否存在超级权限的ID以及正在使用的ID，存在则不能删除
+            if (in_array(1,$array) == false){
+                if (in_array(Session::get('Admin.id'),$array) == false){
+                    //进行删除操作
+                    $res = Db::name('group')->delete($id);
+                    if ($res){
+                        $this->success("删除成功！");
+                    }else{
+                        $this->error("删除失败！");
+                    }
+                }else{
+                    $this->error("正在使用的权限组不能删除！");
+                }
             }else{
-                $this->error("删除失败！");
+                $this->error("超级权限组不能删除！");
             }
         }
     }
